+++ cola-samples/charge/src/main/java/com/huawei/charging/adapter/ChargeController.java
@@ -2,10 +2,10 @@ package com.huawei.charging.adapter;
 
 import com.huawei.charging.application.ChargeServiceI;
 import com.huawei.charging.application.dto.*;
+import jakarta.annotation.Resource;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 
-import javax.annotation.Resource;
 
 @RestController
 @Slf4j
